requirements.txt
# requirements.txt:
# --------------------------------------------------------------------------------
# Filename: requirements.txt
# Description: Python dependencies for the Boutique AI System.
# Version: 4.8 (Ignis Full System Recalibration)

# --- Core Framework & Web ---
fastapi>=0.110.0,<0.111.0 # Pinned to a recent stable minor version
uvicorn[standard]>=0.27.0,<0.28.0 # Pinned, [standard] includes websockets

# --- HTTP Client (Crucial Pin - for Supabase, OpenAI, Deepgram if it used it) ---
# supabase-py 2.x series is generally compatible with more recent httpx
# openai 1.x is flexible with httpx >=0.23.0
# deepgram-sdk 3.x uses its own aiohttp or httpx based on version.
# Let's allow a slightly broader httpx range compatible with recent supabase-py and openai.
httpx>=0.25.2,<1.0.0

# --- AI & LLMs ---
openai>=1.20.0,<2.0.0 # Allowing recent 1.x versions

# --- Database ---
# For supabase-py 2.x, which is generally more up-to-date with httpx
supabase>=2.5.0,<2.16.0 # Targeting a recent stable 2.x range (e.g. from late 2024 to Q1 2025)

# --- External Services ---
deepgram-sdk>=3.7.0,<3.8.0 # This version uses httpx ^0.25.2, compatible with above httpx
twilio>=9.0.0,<9.4.0 # Stable range
websockets>=12.0,<13.0 # Usually fine, uvicorn[standard] can also manage this

# --- Other Core Libraries ---
sqlalchemy>=2.0.25,<2.1.0 # Stable 2.0.x
asyncpg>=0.28.0,<0.30.0 # Recent stable
psutil>=5.9.0,<6.1.0
aiohttp>=3.9.0,<3.10.0 # For DataWrapper and Twilio's aiohttp-retry

# --- Web Interaction & Scraping ---
playwright>=1.40.0,<1.48.0 # Broader stable range
beautifulsoup4>=4.12.0,<4.13.0
Faker>=24.0.0,<27.0.0
fake-useragent>=1.5.0,<2.0.0

# --- Utilities ---
pydantic-settings>=2.1.0,<2.4.0
tenacity>=8.2.0,<9.0.0
pybreaker>=1.1.0,<1.3.0
python-dotenv>=1.0.0,<1.1.0
pytz>=2024.1,<2025.0
cryptography>=42.0.0 # Ensure recent enough for Python 3.11 and security
requests>=2.31.0,<2.33.0

# --- Other Dependencies ---
numpy>=1.26.0,<1.27.0
docker>=7.0.0,<8.0.0 # For DeploymentManager
lru-dict>=1.1.8,<1.3.0 # For LLMClient cache (explicitly add as it's imported)
# --------------------------------------------------------------------------------